第一次提交

+++ gd-wms-outbound-server/src/main/java/com/bpm/gd/wms/outbound/server/domain/BaseDepotToMatter.java
@@ -2,6 +2,7 @@ package com.bpm.gd.wms.outbound.server.domain;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -12,7 +13,7 @@ public class BaseDepotToMatter {
 
     private String matterId;
 
-    private Long allNumber;
+    private BigDecimal allNumber;
 
     private Long inUseNumber;
